[Source] Use int64_t.

+++ main/src/widget/converter.cpp
@@ -47,8 +47,8 @@ namespace khopper {
 				throw error::RunTimeError( "Can not open decoder or encoder!" );
 			}
 
-			int begin = track->get( "start_time" ).value< album::Index >().toMillisecond();
-			int end = begin + track->get( "duration" ).value< album::Index >().toMillisecond();
+			int64_t begin = track->get( "start_time" ).value< album::Index >().toMillisecond();
+			int64_t end = begin + track->get( "duration" ).value< album::Index >().toMillisecond();
 			decoder->setRange( begin, end );
 			if( !decoder->seek( begin ) ) {
 				throw error::CodecError( "Invalid start point" );
